requirements.txt
Flask==2.2.2
Flask-BabelEx==0.9.4
Flask-Login==0.6.2
Flask-RESTful==0.3.9
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
Jinja2==3.1.2
WTForms==3.0.1
